Added the models for Exercise and Workout.

+++ DotnetWebAPI/WorkoutTracker/Exercise/AddExerciseRequest.cs
@@ -4,7 +4,7 @@ public record AddExerciseRequest(
     Guid LinkedWorkoutID,
     string ExerciseName,
     int NumberOfSets,
-    List<int> WeightEachSet,
+    List<float> WeightEachSet,
     List<int> RepsEachSet,
     string notes
 );
