+++ DotnetWebAPI/WorkoutTracker/Exercise/ExerciseResponse.cs
@@ -5,7 +5,7 @@ public record ExerciseResponse(
     Guid ExerciseID,
     string ExerciseName,
     int NumberOfSets,
-    List<int> WeightEachSet,
+    List<float> WeightEachSet,
     List<int> RepsEachSet,
     string notes
 );
