+++ DotnetWebAPI/WorkoutTracker/Exercise/UpsertExerciseRequest.cs
@@ -5,7 +5,7 @@ public record UpsertExerciseRequest(
     Guid ExerciseID,
     string ExerciseName,
     int NumberOfSets,
-    List<int> WeightEachSet,
+    List<float> WeightEachSet,
     List<int> RepsEachSet,
     string notes
 );
